IEEE-319 Converted Navbar.js and Navbar.test.tsx and Updated Cart Quantity in Navbar
